# engine/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Knowledge Graph API",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/notes": {
      "get": {
        "summary": "List Notes",
        "operationId": "list_notes_notes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NoteOut"
                  },
                  "type": "array",
                  "title": "Response List Notes Notes Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Note",
        "operationId": "create_note_notes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/notes/{note_id}": {
      "get": {
        "summary": "Get Note",
        "operationId": "get_note_notes__note_id__get",
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Note Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/notes/{note_id}/find-connections": {
      "post": {
        "summary": "Find Connections",
        "operationId": "find_connections_notes__note_id__find_connections_post",
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Note Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scan": {
      "post": {
        "summary": "Scan",
        "operationId": "scan_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/connections": {
      "get": {
        "summary": "List Connections",
        "operationId": "list_connections_connections_get",
        "parameters": [
          {
            "name": "note_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Note Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConnectionOut"
                  },
                  "title": "Response List Connections Connections Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "summary": "Get Job",
        "operationId": "get_job_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ConnectionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "a_id": {
            "type": "string",
            "title": "A Id"
          },
          "b_id": {
            "type": "string",
            "title": "B Id"
          },
          "a_title": {
            "type": "string",
            "title": "A Title"
          },
          "b_title": {
            "type": "string",
            "title": "B Title"
          },
          "facet_type": {
            "type": "string",
            "title": "Facet Type"
          },
          "kind": {
            "type": "string",
            "enum": [
              "same mechanism",
              "same dynamic",
              "same topic"
            ],
            "title": "Kind"
          },
          "statement": {
            "type": "string",
            "title": "Statement"
          },
          "validity": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Validity"
          },
          "nonobviousness": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Nonobviousness"
          },
          "q": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Q"
          }
        },
        "type": "object",
        "required": [
          "id",
          "a_id",
          "b_id",
          "a_title",
          "b_title",
          "facet_type",
          "kind",
          "statement",
          "validity",
          "nonobviousness",
          "q"
        ],
        "title": "ConnectionOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobOut": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "done",
              "error"
            ],
            "title": "Status"
          },
          "surfaced_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surfaced Count"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "JobOut"
      },
      "NoteCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "body": {
            "type": "string",
            "minLength": 1,
            "title": "Body"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "authored"
          }
        },
        "type": "object",
        "required": [
          "body"
        ],
        "title": "NoteCreate"
      },
      "NoteDetail": {
        "properties": {
          "note": {
            "$ref": "#/components/schemas/NoteOut"
          },
          "connections": {
            "items": {
              "$ref": "#/components/schemas/ConnectionOut"
            },
            "type": "array",
            "title": "Connections"
          }
        },
        "type": "object",
        "required": [
          "note"
        ],
        "title": "NoteDetail"
      },
      "NoteOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "connection_count": {
            "type": "integer",
            "title": "Connection Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "body",
          "source",
          "created_at"
        ],
        "title": "NoteOut"
      },
      "ScanRequest": {
        "properties": {
          "full": {
            "type": "boolean",
            "title": "Full",
            "default": false
          }
        },
        "type": "object",
        "title": "ScanRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}